requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
folium>=0.14.0
python-dotenv>=1.0.0
streamlit-chat>=0.1.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
